add ccsl data

+++ omp2cd-master3/darts-x86/example/GpuTest/StencilCuda3D/StencilCuda3D_7Points.2/StencilTP.h
@@ -142,7 +142,9 @@ DEF_TP(StencilTP)
 	
     uint64_t lastCnt = 0;
     uint64_t gpuWLMax= 0;
-    size_t gpuMemMax = 0;
+    uint64_t gpuMemMax = 0;
+    
+
     bool invokeStreams = false;
     int nStream = 4 ;
 	cudaStream_t *stream ;
